Move CompartmentalModel from tempaltes.h to compartmentalmodel.h
